Increase channelID width in ROOT files to handle larger systems

+++ src/petsys_util/convert_raw_to_raw.cpp
@@ -29,7 +29,7 @@ private:
 	long long 	brStepEnd;
 
 	long long	brFrameID;
-	unsigned short	brChannelID;
+	unsigned int	brChannelID;
 	unsigned short	brTacID;
 	unsigned short	brTCoarse;
 	unsigned short	brECoarse;
